Enable synthesis from phonological feature vectors
Input text is first mapped to phonemes using espeak as before, then IPA
symbols (and remaining punctuation) are converted to SPE-style binary
feature vectors.

Embedding layer is replaced with a 1-D convolution over input features
per timestep. This gives a 512-D context-independent representation of
each phone in the sequence, which is a weighted sum of multi-hot feature
indices. This should be equivalent in principle to an embedding lookup
per input symbol.